pyproject.toml
[project]
name = "compression"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fireworks-ai>=0.19.20",
    "openai>=1.72.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.1.0",
]
